+++ Public/Invoke-FilesToPrompt.ps1
@@ -5,12 +5,12 @@
 .DESCRIPTION
     Takes one or more paths to files or directories and outputs every file, recursively, each one preceded with its filename like this:
 
-    path/to/file.py
+    path/to/file.ps1
     ----
-    Contents of file.py goes here
+    Contents of file.ps1 goes here
 
     ---
-    path/to/file2.py
+    path/to/file2.ps1
     ---
     ...
 
@@ -27,11 +27,15 @@
     
     This example invokes the Invoke-FilesToPrompt function to process all Markdown files in the "C:\MyFiles" directory.
 
-    .EXAMPLE
+.EXAMPLE
     Invoke-FilesToPrompt -Path "C:\MyFiles\*.md", "C:\MyOtherFiles\*.md"
     
     This example invokes the Invoke-FilesToPrompt function to process all Markdown files in the "C:\MyFiles" directory.
 
+.EXAMPLE
+    (Invoke-FilesToPrompt (dir . -r *.md))
+
+    This example recursively processes all Markdown files in the current directory.
 #>
 function Invoke-FilesToPrompt {
     [CmdletBinding()]